studentcms/requirements.txt
black==23.11.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
Django==4.2.7
psycopg2==2.9.9
pyjwt==2.8.0
